change sql to use decimal instead of float as datatype for money related.

+++ api/migrations/m00008_items.php
@@ -14,7 +14,7 @@ class m00008_items extends MigrationScheme
         $sql = "CREATE TABLE IF NOT EXISTS items (
                     item_id int AUTO_INCREMENT NOT NULL PRIMARY KEY,
                     name varchar(255) NOT NULL,
-                    price float,
+                    price decimal(15,3),
                     category_ids varchar(255),
                     blocked bool NOT NULL     
                 )";
